Fix prepare_job_func()
The structure of the function was broken: we did never reset the buffer.
And we tried to initialize the iov without having a clue how much we
actually write to the buffer.

Those two errors ironically made it "work" again... until the buffer was
full.

This commit fixes the issue by moving stuff around and adding a call to
`buf_reset()`.